// src/app/api/models/category.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Category {
  code: string;
  description?: string;
  display_order?: number;
  full_name: string;
  id: number;
  image?: string | null;
  is_active?: boolean;
  name: string;
  name_kh?: string;
  parent?: number | null;
  parent_name: string | null;
}